+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to Toisto will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v0.2.0 - 2022-11-27
+## v0.2.0 - 2022-11-29
+
+### Fixed
+
+- When showing the meaning of a concept, show only the first spelling variant. Fixes [#63](https://github.com/fniessink/toisto/issues/63).
 
 ### Added